+++ psd-web/app/decorators/corrective_action_decorator.rb
@@ -1,6 +1,7 @@
 class CorrectiveActionDecorator < ApplicationDecorator
   delegate_all
   include SupportingInformationHelper
+  include SupportingInformation::CorrectiveActionSortInterface
 
   def details
     return if object.details.blank?
@@ -13,11 +14,7 @@ class CorrectiveActionDecorator < ApplicationDecorator
   end
 
   def date_of_activity
-    date_decided
-  end
-
-  def date_of_activity_string
-    date_of_activity.to_s(:govuk)
+    date_decided.to_s(:govuk)
   end
 
   def date_added
