+++ psd-web/app/models/concerns/supporting_information/corrective_action_sort_interface.rb
@@ -1,0 +1,9 @@
+module SupportingInformation
+  module CorrectiveActionSortInterface
+    extend ActiveSupport::Concern
+
+    def date_of_activity_sort
+      date_decided
+    end
+  end
+end
