+++ psd-web/app/models/concerns/supporting_information/correspondence_sort_interface.rb
@@ -1,0 +1,9 @@
+module SupportingInformation
+  module CorrespondenceSortInterface
+    extend ActiveSupport::Concern
+
+    def date_of_activity_sort
+      correspondence_date
+    end
+  end
+end
