+++ psd-web/app/models/concerns/supporting_information/test_result_sort_interface.rb
@@ -1,0 +1,9 @@
+module SupportingInformation
+  module TestResultSortInterface
+    extend ActiveSupport::Concern
+
+    def date_of_activity_sort
+      date
+    end
+  end
+end
