+++ psd-web/app/views/investigations/supporting_information/_supporting_information_row.erb
@@ -6,7 +6,7 @@
     <%= supporting_information_row.supporting_information_type %>
   </td>
   <td class="govuk-table__cell">
-    <%= supporting_information_row.date_of_activity_string %>
+    <%= supporting_information_row.date_of_activity %>
   </td>
   <td class="govuk-table__cell">
     <%= supporting_information_row.date_added %>
