+++ psd-web/spec/helpers/investigations/corrective_actions_helper_spec.rb
@@ -7,7 +7,7 @@ RSpec.describe Investigations::CorrectiveActionsHelper, :with_stubbed_elasticsea
 
     let(:expected_rows) do
       [
-        { key: { text: "Date of action" }, value: { text: corrective_action.date_of_activity_string } },
+        { key: { text: "Date of action" }, value: { text: corrective_action.date_of_activity } },
         { key: { text: "Legislation" }, value: { text: corrective_action.legislation } },
         { key: { text: "Product" }, value: { html: helper.link_to(corrective_action.product.name, helper.product_path(corrective_action.product)) } },
         { key: { text: "Business" }, value: { html: helper.link_to(corrective_action.business.trading_name, helper.business_path(corrective_action.business)) } },
